Update Q9.cpp
Divides the runtime by 17. The big O goes from n^3 to n ^2, while increasing the processing time in each loop. Instead of looping through what C could be, I am defining C as from a and b. It is clearly is worth it when you compare the runtimes. Still worth investigating the order of the if statement on line 19.